Separate building from packaging
